Added checking for file existence.

+++ test.py
@@ -47,8 +47,9 @@ for title_data in titles_data:
 word_avg = word_sum / n
 char_avg = char_sum / n
 
-with open("frequencies.json", "w") as write_file:
-    write_file.write(json.dumps(chars))
+if os.path.isfile("frequencies.json"):
+    with open("frequencies.json", "w") as write_file:
+        write_file.write(json.dumps(chars))
 
 print "Word length Average: %s" % word_avg
 print "Character length Average: %s" % char_avg
